refactor: split worktree files into individual method files
Split worktree.go and worktree_test.go into individual files for better organization and maintainability. Each method from the Worktree interface now has its own dedicated file with corresponding tests.

Changes:
- Split worktree.go into individual method files (build_path.go, create.go, delete.go, etc.)
- Split worktree_test.go into corresponding test files for each method
- Created helpers.go for internal helper methods
- Updated worktree.go to contain only type definitions and imports
- Maintained all existing functionality and test coverage
- All 24 unit tests pass
- No linting errors
- All integration and e2e tests pass

This refactoring improves code organization by having one method per file, making it easier to navigate and maintain the worktree package.
