+++ pkg/worktree/build_path_test.go
@@ -1,0 +1,38 @@
+//go:build unit
+
+package worktree
+
+import (
+	"testing"
+
+	fsmocks "github.com/lerenn/code-manager/pkg/fs/mocks"
+	gitmocks "github.com/lerenn/code-manager/pkg/git/mocks"
+	"github.com/lerenn/code-manager/pkg/logger"
+	promptmocks "github.com/lerenn/code-manager/pkg/prompt/mocks"
+	statusmocks "github.com/lerenn/code-manager/pkg/status/mocks"
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
+)
+
+func TestWorktree_BuildPath(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	mockFS := fsmocks.NewMockFS(ctrl)
+	mockGit := gitmocks.NewMockGit(ctrl)
+	mockStatus := statusmocks.NewMockManager(ctrl)
+	mockPrompt := promptmocks.NewMockPrompter(ctrl)
+
+	worktree := &realWorktree{
+		fs:              mockFS,
+		git:             mockGit,
+		statusManager:   mockStatus,
+		logger:          logger.NewNoopLogger(),
+		prompt:          mockPrompt,
+		repositoriesDir: "/test/base",
+	}
+
+	path := worktree.BuildPath("github.com/octocat/Hello-World", "origin", "feature-branch")
+	expected := "/test/base/github.com/octocat/Hello-World/origin/feature-branch"
+	assert.Equal(t, expected, path)
+}
