+++ pkg/worktree/exists.go
@@ -1,0 +1,7 @@
+// Package worktree provides worktree management functionality for CM.
+package worktree
+
+// Exists checks if a worktree exists for the specified branch.
+func (w *realWorktree) Exists(repoPath, branch string) (bool, error) {
+	return w.git.WorktreeExists(repoPath, branch)
+}
