+++ pkg/worktree/validate_creation_test.go
@@ -1,0 +1,48 @@
+//go:build unit
+
+package worktree
+
+import (
+	"errors"
+	"testing"
+
+	fsmocks "github.com/lerenn/code-manager/pkg/fs/mocks"
+	gitmocks "github.com/lerenn/code-manager/pkg/git/mocks"
+	promptmocks "github.com/lerenn/code-manager/pkg/prompt/mocks"
+	statusmocks "github.com/lerenn/code-manager/pkg/status/mocks"
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
+)
+
+func TestWorktree_ValidateCreation_Success(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	mockFS := fsmocks.NewMockFS(ctrl)
+	mockGit := gitmocks.NewMockGit(ctrl)
+	mockStatus := statusmocks.NewMockManager(ctrl)
+	mockPrompt := promptmocks.NewMockPrompter(ctrl)
+
+	worktree := &realWorktree{
+		fs:              mockFS,
+		git:             mockGit,
+		statusManager:   mockStatus,
+		prompt:          mockPrompt,
+		repositoriesDir: "/test/base",
+	}
+
+	params := ValidateCreationParams{
+		RepoURL:      "github.com/octocat/Hello-World",
+		Branch:       "feature-branch",
+		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
+		RepoPath:     "/test/repo",
+	}
+
+	// Mock expectations
+	mockFS.EXPECT().Exists(params.WorktreePath).Return(false, nil)
+	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(nil, errors.New("not found"))
+	mockFS.EXPECT().MkdirAll(gomock.Any(), gomock.Any()).Return(nil)
+
+	err := worktree.ValidateCreation(params)
+	assert.NoError(t, err)
+}
