--- pkg/worktree/worktree_test.go
@@ -1,783 +1,0 @@
-//go:build unit
-
-package worktree
-
-import (
-	"errors"
-	"testing"
-
-	fsmocks "github.com/lerenn/code-manager/pkg/fs/mocks"
-	"github.com/lerenn/code-manager/pkg/git"
-	gitmocks "github.com/lerenn/code-manager/pkg/git/mocks"
-	"github.com/lerenn/code-manager/pkg/logger"
-	promptmocks "github.com/lerenn/code-manager/pkg/prompt/mocks"
-	"github.com/lerenn/code-manager/pkg/status"
-	statusmocks "github.com/lerenn/code-manager/pkg/status/mocks"
-	"github.com/stretchr/testify/assert"
-	"go.uber.org/mock/gomock"
-)
-
-func TestNewWorktree(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	params := NewWorktreeParams{
-		FS:              mockFS,
-		Git:             mockGit,
-		StatusManager:   mockStatus,
-		Prompt:          mockPrompt,
-		RepositoriesDir: "/test/path",
-	}
-
-	worktree := NewWorktree(params)
-	assert.NotNil(t, worktree)
-}
-
-func TestWorktree_BuildPath(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	path := worktree.BuildPath("github.com/octocat/Hello-World", "origin", "feature-branch")
-	expected := "/test/base/github.com/octocat/Hello-World/origin/feature-branch"
-	assert.Equal(t, expected, path)
-}
-
-func TestWorktree_Create_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs: mockFS, git: mockGit, statusManager: mockStatus, logger: logger.NewNoopLogger(), prompt: mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := CreateParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Remote:       "origin",
-		IssueInfo:    nil,
-		Force:        false,
-	}
-
-	// Mock expectations
-	mockFS.EXPECT().Exists(params.WorktreePath).Return(false, nil)
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(nil, errors.New("not found"))
-	mockFS.EXPECT().MkdirAll(gomock.Any(), gomock.Any()).Return(nil)
-	mockGit.EXPECT().CheckReferenceConflict(params.RepoPath, params.Branch).Return(nil)
-	mockGit.EXPECT().BranchExists(params.RepoPath, params.Branch).Return(true, nil)
-	mockFS.EXPECT().MkdirAll(params.WorktreePath, gomock.Any()).Return(nil)
-	mockGit.EXPECT().CreateWorktreeWithNoCheckout(params.RepoPath, params.WorktreePath, params.Branch).Return(nil)
-
-	err := worktree.Create(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_Create_DirectoryExists(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs: mockFS, git: mockGit, statusManager: mockStatus, prompt: mockPrompt, repositoriesDir: "/test/base",
-	}
-
-	params := CreateParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Remote:       "origin",
-		IssueInfo:    nil,
-		Force:        false,
-	}
-
-	// Mock expectations
-	mockFS.EXPECT().Exists(params.WorktreePath).Return(true, nil)
-
-	err := worktree.Create(params)
-	assert.ErrorIs(t, err, ErrDirectoryExists)
-}
-
-func TestWorktree_Create_WorktreeExists(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := CreateParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Remote:       "origin",
-		IssueInfo:    nil,
-		Force:        false,
-	}
-
-	existingWorktree := &status.WorktreeInfo{
-		Branch: params.Branch,
-		Remote: "origin",
-	}
-
-	// Mock expectations
-	mockFS.EXPECT().Exists(params.WorktreePath).Return(false, nil)
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(existingWorktree, nil)
-
-	err := worktree.Create(params)
-	assert.ErrorIs(t, err, ErrWorktreeExists)
-}
-
-func TestWorktree_Create_BranchDoesNotExist(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := CreateParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Remote:       "origin",
-		IssueInfo:    nil,
-		Force:        false,
-	}
-
-	// Mock expectations
-	mockFS.EXPECT().Exists(params.WorktreePath).Return(false, nil)
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(nil, errors.New("not found"))
-	mockFS.EXPECT().MkdirAll(gomock.Any(), gomock.Any()).Return(nil)
-	mockGit.EXPECT().CheckReferenceConflict(params.RepoPath, params.Branch).Return(nil)
-	mockGit.EXPECT().BranchExists(params.RepoPath, params.Branch).Return(false, nil)
-	mockGit.EXPECT().FetchRemote(params.RepoPath, "origin").Return(nil)
-	mockGit.EXPECT().BranchExistsOnRemote(git.BranchExistsOnRemoteParams{
-		RepoPath:   params.RepoPath,
-		RemoteName: "origin",
-		Branch:     params.Branch,
-	}).Return(false, nil)
-	mockGit.EXPECT().GetRemoteURL(params.RepoPath, "origin").Return("https://github.com/octocat/Hello-World.git", nil)
-	mockGit.EXPECT().GetDefaultBranch("https://github.com/octocat/Hello-World.git").Return("main", nil)
-	mockGit.EXPECT().CreateBranchFrom(git.CreateBranchFromParams{
-		RepoPath:   params.RepoPath,
-		NewBranch:  params.Branch,
-		FromBranch: "origin/main",
-	}).Return(nil)
-	mockFS.EXPECT().MkdirAll(params.WorktreePath, gomock.Any()).Return(nil)
-	mockGit.EXPECT().CreateWorktreeWithNoCheckout(params.RepoPath, params.WorktreePath, params.Branch).Return(nil)
-
-	err := worktree.Create(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_Delete_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := DeleteParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Force:        true,
-	}
-
-	existingWorktree := &status.WorktreeInfo{
-		Branch: params.Branch,
-		Remote: "origin",
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(existingWorktree, nil)
-	mockGit.EXPECT().RemoveWorktree(params.RepoPath, params.WorktreePath, params.Force).Return(nil)
-	mockFS.EXPECT().RemoveAll(params.WorktreePath).Return(nil)
-	mockStatus.EXPECT().RemoveWorktree(params.RepoURL, params.Branch).Return(nil)
-
-	err := worktree.Delete(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_Delete_WorktreeNotInStatus(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := DeleteParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Force:        true,
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(nil, errors.New("not found"))
-
-	err := worktree.Delete(params)
-	assert.ErrorIs(t, err, ErrWorktreeNotInStatus)
-}
-
-func TestWorktree_Delete_WithConfirmation(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := DeleteParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Force:        false,
-	}
-
-	existingWorktree := &status.WorktreeInfo{
-		Branch: params.Branch,
-		Remote: "origin",
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(existingWorktree, nil)
-	mockPrompt.EXPECT().PromptForConfirmation(gomock.Any(), false).Return(true, nil)
-	mockGit.EXPECT().RemoveWorktree(params.RepoPath, params.WorktreePath, params.Force).Return(nil)
-	mockFS.EXPECT().RemoveAll(params.WorktreePath).Return(nil)
-	mockStatus.EXPECT().RemoveWorktree(params.RepoURL, params.Branch).Return(nil)
-
-	err := worktree.Delete(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_Delete_ConfirmationCancelled(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := DeleteParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-		Force:        false,
-	}
-
-	existingWorktree := &status.WorktreeInfo{
-		Branch: params.Branch,
-		Remote: "origin",
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(existingWorktree, nil)
-	mockPrompt.EXPECT().PromptForConfirmation(gomock.Any(), false).Return(false, nil)
-
-	err := worktree.Delete(params)
-	assert.ErrorIs(t, err, ErrDeletionCancelled)
-}
-
-func TestWorktree_ValidateCreation_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := ValidateCreationParams{
-		RepoURL:      "github.com/octocat/Hello-World",
-		Branch:       "feature-branch",
-		WorktreePath: "/test/base/github.com/octocat/Hello-World/origin/feature-branch",
-		RepoPath:     "/test/repo",
-	}
-
-	// Mock expectations
-	mockFS.EXPECT().Exists(params.WorktreePath).Return(false, nil)
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(nil, errors.New("not found"))
-	mockFS.EXPECT().MkdirAll(gomock.Any(), gomock.Any()).Return(nil)
-
-	err := worktree.ValidateCreation(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_ValidateDeletion_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := ValidateDeletionParams{
-		RepoURL: "github.com/octocat/Hello-World",
-		Branch:  "feature-branch",
-	}
-
-	existingWorktree := &status.WorktreeInfo{
-		Branch: params.Branch,
-		Remote: "origin",
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().GetWorktree(params.RepoURL, params.Branch).Return(existingWorktree, nil)
-
-	err := worktree.ValidateDeletion(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_EnsureBranchExists_BranchExists(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations
-	mockGit.EXPECT().CheckReferenceConflict(repoPath, branch).Return(nil)
-	mockGit.EXPECT().BranchExists(repoPath, branch).Return(true, nil)
-
-	err := worktree.EnsureBranchExists(repoPath, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_EnsureBranchExists_BranchDoesNotExist(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations
-	mockGit.EXPECT().CheckReferenceConflict(repoPath, branch).Return(nil)
-	mockGit.EXPECT().BranchExists(repoPath, branch).Return(false, nil)
-	mockGit.EXPECT().FetchRemote(repoPath, "origin").Return(nil)
-	mockGit.EXPECT().BranchExistsOnRemote(git.BranchExistsOnRemoteParams{
-		RepoPath:   repoPath,
-		RemoteName: "origin",
-		Branch:     branch,
-	}).Return(false, nil)
-	mockGit.EXPECT().GetRemoteURL(repoPath, "origin").Return("https://github.com/octocat/Hello-World.git", nil)
-	mockGit.EXPECT().GetDefaultBranch("https://github.com/octocat/Hello-World.git").Return("main", nil)
-	mockGit.EXPECT().CreateBranchFrom(git.CreateBranchFromParams{
-		RepoPath:   repoPath,
-		NewBranch:  branch,
-		FromBranch: "origin/main",
-	}).Return(nil)
-
-	err := worktree.EnsureBranchExists(repoPath, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_EnsureBranchExists_BranchExistsOnRemote(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations
-	mockGit.EXPECT().CheckReferenceConflict(repoPath, branch).Return(nil)
-	mockGit.EXPECT().BranchExists(repoPath, branch).Return(false, nil)
-	mockGit.EXPECT().FetchRemote(repoPath, "origin").Return(nil)
-	mockGit.EXPECT().BranchExistsOnRemote(git.BranchExistsOnRemoteParams{
-		RepoPath:   repoPath,
-		RemoteName: "origin",
-		Branch:     branch,
-	}).Return(true, nil)
-	mockGit.EXPECT().CreateBranchFrom(git.CreateBranchFromParams{
-		RepoPath:   repoPath,
-		NewBranch:  branch,
-		FromBranch: "origin/" + branch,
-	}).Return(nil)
-
-	err := worktree.EnsureBranchExists(repoPath, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_EnsureBranchExists_BranchDoesNotExist_RemoteFallback(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations - remote operations fail, fallback to local
-	mockGit.EXPECT().CheckReferenceConflict(repoPath, branch).Return(nil)
-	mockGit.EXPECT().BranchExists(repoPath, branch).Return(false, nil)
-	mockGit.EXPECT().FetchRemote(repoPath, "origin").Return(nil)
-	mockGit.EXPECT().BranchExistsOnRemote(git.BranchExistsOnRemoteParams{
-		RepoPath:   repoPath,
-		RemoteName: "origin",
-		Branch:     branch,
-	}).Return(false, nil)
-	mockGit.EXPECT().GetRemoteURL(repoPath, "origin").Return("", errors.New("no remote"))
-	mockGit.EXPECT().GetCurrentBranch(repoPath).Return("main", nil)
-	mockGit.EXPECT().CreateBranchFrom(git.CreateBranchFromParams{
-		RepoPath:   repoPath,
-		NewBranch:  branch,
-		FromBranch: "main",
-	}).Return(nil)
-
-	err := worktree.EnsureBranchExists(repoPath, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_EnsureBranchExists_BranchDoesNotExist_DefaultBranchFallback(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:              mockFS,
-		git:             mockGit,
-		statusManager:   mockStatus,
-		logger:          logger.NewNoopLogger(),
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations - remote exists but default branch detection fails, fallback to local
-	mockGit.EXPECT().CheckReferenceConflict(repoPath, branch).Return(nil)
-	mockGit.EXPECT().BranchExists(repoPath, branch).Return(false, nil)
-	mockGit.EXPECT().FetchRemote(repoPath, "origin").Return(nil)
-	mockGit.EXPECT().BranchExistsOnRemote(git.BranchExistsOnRemoteParams{
-		RepoPath:   repoPath,
-		RemoteName: "origin",
-		Branch:     branch,
-	}).Return(false, nil)
-	mockGit.EXPECT().GetRemoteURL(repoPath, "origin").Return("https://github.com/octocat/Hello-World.git", nil)
-	mockGit.EXPECT().GetDefaultBranch("https://github.com/octocat/Hello-World.git").Return("", errors.New("failed to get default branch"))
-	mockGit.EXPECT().GetCurrentBranch(repoPath).Return("main", nil)
-	mockGit.EXPECT().CreateBranchFrom(git.CreateBranchFromParams{
-		RepoPath:   repoPath,
-		NewBranch:  branch,
-		FromBranch: "main",
-	}).Return(nil)
-
-	err := worktree.EnsureBranchExists(repoPath, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_AddToStatus_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	params := AddToStatusParams{
-		RepoURL:       "github.com/octocat/Hello-World",
-		Branch:        "feature-branch",
-		WorktreePath:  "/test/path",
-		WorkspacePath: "",
-		Remote:        "origin",
-		IssueInfo:     nil,
-	}
-
-	// Mock expectations
-	mockStatus.EXPECT().AddWorktree(gomock.Any()).Return(nil)
-
-	err := worktree.AddToStatus(params)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_RemoveFromStatus_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoURL := "github.com/octocat/Hello-World"
-	branch := "feature-branch"
-
-	// Mock expectations
-	mockStatus.EXPECT().RemoveWorktree(repoURL, branch).Return(nil)
-
-	err := worktree.RemoveFromStatus(repoURL, branch)
-	assert.NoError(t, err)
-}
-
-func TestWorktree_Exists_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	repoPath := "/test/repo"
-	branch := "feature-branch"
-
-	// Mock expectations
-	mockGit.EXPECT().WorktreeExists(repoPath, branch).Return(true, nil)
-
-	exists, err := worktree.Exists(repoPath, branch)
-	assert.NoError(t, err)
-	assert.True(t, exists)
-}
-
-func TestWorktree_CleanupDirectory_Success(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockFS := fsmocks.NewMockFS(ctrl)
-	mockGit := gitmocks.NewMockGit(ctrl)
-	mockStatus := statusmocks.NewMockManager(ctrl)
-
-	mockPrompt := promptmocks.NewMockPrompter(ctrl)
-
-	worktree := &realWorktree{
-		fs:            mockFS,
-		git:           mockGit,
-		statusManager: mockStatus,
-
-		prompt:          mockPrompt,
-		repositoriesDir: "/test/base",
-	}
-
-	worktreePath := "/test/worktree/path"
-
-	// Mock expectations
-	mockFS.EXPECT().RemoveAll(worktreePath).Return(nil)
-
-	err := worktree.CleanupDirectory(worktreePath)
-	assert.NoError(t, err)
-}
